+++ srcs/raycaster/setup_frame.c
@@ -6,14 +6,14 @@
 /*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/05 10:26:00 by jnivala           #+#    #+#             */
-/*   Updated: 2021/02/19 10:26:29 by jnivala          ###   ########.fr       */
+/*   Updated: 2021/02/19 13:17:22 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../doom_nukem.h"
 
 void		setup_frame(t_frame *frame, t_frame *new_frame,
-	float current_angle, int idx)
+	double current_angle, int idx)
 {
 	new_frame->idx = idx;
 	new_frame->old_idx = frame->idx;
